iraqi_aware/requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
Pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0